// src/generated/models/Cfm_cashflowcategoriesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cfm_cashflowcategoriesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cfm_cashflowcategoriesstatecode = keyof typeof Cfm_cashflowcategoriesstatecode;
export const Cfm_cashflowcategoriesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cfm_cashflowcategoriesstatuscode = keyof typeof Cfm_cashflowcategoriesstatuscode;

export interface Cfm_cashflowcategoriesBase {
  cfm_amount?: number;
  cfm_cash_flow_category?: string;
  cfm_cashdirection?: string;
  cfm_cashflowactivity?: string;
  cfm_cashflowcategoryid: string;
  cfm_date?: string;
  cfm_datekey?: number;
  cfm_name: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cfm_cashflowcategoriesstatecode;
  statuscode?: Cfm_cashflowcategoriesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cfm_cashflowcategories extends Cfm_cashflowcategoriesBase {
  cfm_key?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}