// src/generated/models/Cfm_cashflownetmeasuresesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Cfm_cashflownetmeasuresesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Cfm_cashflownetmeasuresesstatecode = keyof typeof Cfm_cashflownetmeasuresesstatecode;
export const Cfm_cashflownetmeasuresesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Cfm_cashflownetmeasuresesstatuscode = keyof typeof Cfm_cashflownetmeasuresesstatuscode;

export interface Cfm_cashflownetmeasuresesBase {
  cfm_beginningbalance?: number;
  cfm_cashflownetmeasuresid: string;
  cfm_endingbalancecf?: number;
  cfm_financingactivitiesin?: number;
  cfm_financingactivitiesout?: number;
  cfm_investingactivitiesin?: number;
  cfm_investingactivitiesout?: number;
  cfm_month?: string;
  cfm_name: string;
  cfm_netfinancingactivities?: number;
  cfm_netinvestingactivities?: number;
  cfm_netoperationin?: number;
  cfm_netoperationout?: number;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Cfm_cashflownetmeasuresesstatecode;
  statuscode?: Cfm_cashflownetmeasuresesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Cfm_cashflownetmeasureses extends Cfm_cashflownetmeasuresesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}